requirements.txt
anyio==3.5.0
astroid==2.9.3
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
Authlib==0.15.5
autopep8==1.6.0
backcall==0.2.0
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
colorful==0.5.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.2
flatten-json==0.1.13
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
ipykernel==6.9.0
ipython==8.0.1
isort==5.10.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
outcome==1.1.0
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
prettyprinter==0.18.0
prompt-toolkit==3.0.28
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pylint==2.12.2
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
rfc3986==1.5.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
stack-data==0.1.4
tda-api==1.5.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.1.0
urllib3==1.26.8
wcwidth==0.2.5
webdriver-manager==3.5.2
websockets==10.1
wrapt==1.13.3
wsproto==1.0.0
yarg==0.1.9
